separacion de logica y cambio de componente logo
