+++ src/pages/ForgotPassword.tsx
@@ -1,64 +1,16 @@
-import { useState } from 'react';
 import { GradientButton } from '../components/Password/GradientButton';
-import { Logo } from '../components/Password/Logo';
-import clienteAxios from '../config/axios';
+import { Techone } from '../components/register/Techone';
 import { Alert } from '../components/Password/Alert';
-import axios from 'axios';
+import { useForgotPassword } from '../hooks/useForgotPassword';
 
 export const ForgotPassword = () => {
-    const [email, setEmail] = useState('');
-    const [alert, setAlert] = useState<{ mensaje: string; error?: boolean }>({ mensaje: '' });
-
-    const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
-        e.preventDefault();
-
-        if (email === '' || email.length < 6) {
-            setAlert({ mensaje: 'El email es obligatorio', error: true });
-            setTimeout(() => {
-                setAlert({ mensaje: '' });
-            }, 3000);
-            return;
-        }
-
-
-        try {
-            const { data } = await clienteAxios.post('/', { email });
-            setAlert({ mensaje: data.msg, error: false });
-            setTimeout(() => {
-                setAlert({ mensaje: '' });
-            }, 5000);
-        } catch (error) {
-            if (axios.isAxiosError(error) && error.response) {
-                switch (error.response.status) {
-                    case 400:
-                        setAlert({ mensaje: 'Solicitud inválida. Por favor, revisa el email ingresado.', error: true });
-                        break;
-                    case 404:
-                        setAlert({ mensaje: 'No se encontró ninguna cuenta con ese email.', error: true });
-                        break;
-                    case 500:
-                        setAlert({ mensaje: 'Error del servidor. Por favor, intenta de nuevo más tarde.', error: true });
-                        break;
-                    default:
-                        setAlert({ mensaje: error.response.data.msg || 'ha ocurrido un error.', error: true });
-                        break;
-                }
-            } else {
-                setAlert({ mensaje: 'ha ocurrido un error desconocido. Por favor, intenta de nuevo más tarde.', error: true });
-            }
-            setTimeout(() => {
-                setAlert({ mensaje: '' });
-            }, 5000);
-        }
-    };
-
+    const { email, setEmail, alert, handleSubmit } = useForgotPassword();
     const { mensaje } = alert;
-    
 
     return (
         <>
             <div className="flex flex-col justify-center items-center min-h-screen w-full p-4 animate-blurred-fade-in">
-                <Logo />
+                <Techone name="TECHONE" />
                 <h1 className="text-3xl md:text-5xl font-bold text-center text-white mb-6 font-josefin">
                     ¡Vamos a reestablecer tu contraseña!
                 </h1>
