+++ src/pages/NewPassword.tsx
@@ -1,66 +1,21 @@
-import { useState, useEffect } from 'react';
 import { useParams, Link } from 'react-router-dom';
 import { GradientButton } from '../components/Password/GradientButton';
-import { Logo } from '../components/Password/Logo';
-import clienteAxios from '../config/axios';
+import { Techone } from '../components/register/Techone';
 import { Alert } from '../components/Password/Alert';
-
-interface AlertType {
-    mensaje: string;
-    error: boolean;
-}
+import { useTokenValidation } from '../hooks/useTokenValidation';
+import { usePasswordReset } from '../hooks/usePasswordReset';
 
 export const NewPassword = () => {
-    const [password, setPassword] = useState('');
-    const [confirmPassword, setConfirmPassword] = useState('');
-    const [alert, setAlert] = useState<AlertType>({ mensaje: '', error: false });
-    const [tokenValido, setTokenValido] = useState(false);
-    const [passwordModificado, setPasswordModificado] = useState(false);
-
     const { token } = useParams();
-
-    useEffect(() => {
-        const comprobarToken = async () => {
-            try {
-                await clienteAxios(`/${token}`);
-                setAlert({ mensaje: '', error: false });
-                setTokenValido(true);
-            } catch (error) {
-                setAlert({ mensaje: 'Error en el enlace', error: true });
-            }
-        };
-        comprobarToken();
-    }, [token]);
-
-    const handleSubmit = async (e: React.FormEvent) => {
-        e.preventDefault();
-
-        if (password.length < 6) {
-            setAlert({ mensaje: 'La contraseña debe tener más de 6 caracteres', error: true });
-            return;
-        }
-
-        if (password !== confirmPassword) {
-            setAlert({ mensaje: 'Las contraseñas no coinciden', error: true });
-            return;
-        }
-
-        try {
-            const url = `/${token}`;
-            const { data } = await clienteAxios.post(url, { password });
-            setAlert({ mensaje: data.msg, error: false });
-            setPasswordModificado(true);
-        } catch (error: any) {
-            setAlert({ mensaje: error.response?.data?.msg || 'ha ocurrido un error', error: true });
-        }
-    };
+    const { tokenValido, alert, setAlert } = useTokenValidation(token!);
+    const { password, setPassword, confirmPassword, setConfirmPassword, handleSubmit, passwordModificado } = usePasswordReset(token!, setAlert);
 
     const { mensaje } = alert;
 
     return (
         <>
             <div className="flex flex-col justify-center items-center min-h-screen w-full p-4 animate-blurred-fade-in">
-                <Logo />
+                <Techone name="TECHONE" />
                 <h1 className="text-3xl md:text-5xl font-bold text-center text-white mb-6 font-josefin">
                     Ingresa tu nueva contraseña e inicia sesión
                 </h1>
